Finished Ugg & Wrongway, added Coily
Ugg & Wrongway can correctly kill the player.

Coily is added but can only be controlled by the player.

+++ Himinn_Engine/QBert/GridComponent.cpp
@@ -262,7 +262,7 @@ int GridComponent::GetListIndex(int layer, int number) const
 	for (int i = 1; i <= layer; ++i)
 		index += i;
 
-	if (index >= m_pNodes.size())
+	if (index >= (int)m_pNodes.size())
 		return -1;
 	
 	return (index + number);
